requirements.txt
meval==2.5
motor[srv]==3.6.0
pyrofork
tgcrypto-pyrofork
uvloop; sys_platform != "win32"
python-dotenv
